Add title validation to delete confirmation route

+++ routes/index.js
@@ -80,9 +80,12 @@ router.route("/books/new")
             // Redirect to book detail page
             res.redirect(`/books/${newBook.id}`);
 
-        } catch (errors) {
+        } catch (error) {
+            // Set status to 400 Bad Request
+            res.status(400);
+
             // Attach errors to view locals
-            res.locals.errors = errors;
+            res.locals.errors = error.errors;
 
             // Rerender new book form
             res.render("new-book");
@@ -110,6 +113,9 @@ router.route("/books/:id")
             // Redirect to book detail page
             res.redirect(`/books/${updatedBook.id}`);
         } catch (errors) {
+            // Set status to 400 Bad Request
+            res.status(400);
+
             // Attach errors to view locals
             res.locals.errors = errors;
 
@@ -127,11 +133,29 @@ router.route("/books/:id/delete")
         // Render delete confirmation view
         res.render("delete-confirm");
     }).post(asyncHandler(async (req, res) => {
-        // Delete book
-        await req.bookService.delete(req.book);
+        // Ensure that title from form data matches book title
+        const titlesMatch = req.book.title === req.body.title;
 
-        // Redirect to book listing
-        res.redirect("/books");
+        // If the titles match,
+        if (titlesMatch) {
+            // Delete book
+            await req.bookService.delete(req.book);
+
+            // Redirect to book listing
+            res.redirect("/books");
+        } else {
+            // Otherwise, set status to 400 Bad Request
+            res.status(400);
+
+            // Attach error to view locals
+            res.locals.error = new Error("Request body title does not match book title.");
+
+            // Attach book to view locals
+            res.locals.book = req.book;
+
+            // Re-render delete confirmation view
+            res.render("delete-confirm");
+        }
     }));
 
 // Export
